no more My/Their message dirs
