--- src/ChatFeed/MessageList/Message/MyMessage/props.tsx
@@ -1,9 +1,0 @@
-import { MessageProps, ChatProps } from '../../../../util/interfaces';
-
-export interface Props {
-  lastMessage?: MessageProps | null;
-  message: MessageProps;
-  nextMessage?: MessageProps | null;
-  chat?: ChatProps | null;
-  isSending?: boolean;
-}
